feat(front): turn the mobile bottom bar into the home page's tab bar
On the home page the bar switches the sections of that page (menu, chat,
settings) instead of switching places, so the tabs row above the content is
gone. Search keeps its position across both variants.
